timekeeping/requirements.txt
asgiref==3.4.1
Django==3.2.5
django-phonenumber-field==5.2.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
